feat: post authenticate request to BE befoer letting in user to /home
