+++ src/app/[locale]/(main)/layout.tsx
@@ -1,21 +1,38 @@
 'use client';
 
-import { useLocale } from 'next-intl';
+import { useLocale, useTranslations } from 'next-intl';
 import React from 'react';
-import { useAccount } from '~/hooks/account';
+import { HashLoader } from 'react-spinners';
+import { useAuthentication } from '~/hooks/account';
+import { useDotsString } from '~/hooks/loading';
 import { redirect } from '~/i18n/routing';
 import StompProvider from '~/providers/StompProvider';
-import { AccountStatus } from '~/types/account';
 
 const MainLayout: React.FC<React.PropsWithChildren> = ({ children }) => {
-  const { accountStatus } = useAccount();
   const locale = useLocale();
+  const { isError, isPending, isSuccess } = useAuthentication();
+  const t = useTranslations('rootRoute');
+  const dots = useDotsString(3);
 
-  if (accountStatus === AccountStatus.SignedOut) {
+  if (isError) {
     redirect({ locale, href: '/account/sign-in' });
   }
 
-  return <StompProvider>{children}</StompProvider>;
+  return (
+    <>
+      {isPending && (
+        <div className="flex h-screen flex-col items-center justify-center gap-y-4 font-semibold">
+          <HashLoader />
+          <div className="relative">
+            <span>{t('loading')}</span>
+            <span className="absolute">{dots}</span>
+          </div>
+        </div>
+      )}
+
+      {isSuccess && <StompProvider>{children}</StompProvider>}
+    </>
+  );
 };
 
 export default MainLayout;
